pyproject.toml
[tool.poetry]
name = "xj-checkin"
version = "0.1.0"
description = ""
authors = ["KidoThunder <zgdisgod@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
requests = "^2.25.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
